Improve MIME performance lookup
Currently when searching for MIME types to service, a sequencial search is made, wasting time and calls,
specially for the most commonly used types.

This patch introduces a heuristic which allows two mime lists to exist, one with the most commonly used
types (like HTML, CSS, images, etc.), and another with the rest. When servicing a type that is very used, it
will continue to do a sequencial search, otherwise we do a binary one.

Signed-off-by: Davidlohr Bueso <dave@gnu.org>