mk_file_get_info() now accept two parameters and return int value.
the new definition is as follow:

  int mk_file_get_into(const char *path, struct file_info *f_info);

so the called needs to pass by reference an allocated struct for
f_info where the file information will be stored.

At Monkey core level, this new method reduce one malloc/free in the
session_request structure, as the new file_info node is allocated
as static and not dynamic as it used to.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>