Scheduler: export sched node to worker scope through thread_key
Inside a worker, they are multiple calls to the function
mk_sched_get_thread_conf() which returns the sched node associated
to the worker in question. Previously this function was matching
each scheduler node information with the thread ID, now the
function knows the sched node through a thread_key avoiding
'for' cycles and multiple comparissons.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>