Scheduler: store in_addr info instead of IPv4 string
When a connection arrives, the scheduler assign the new connection
to a sched node which hold information about the connection. Previous
this patch, it used to store the IP address in a common string, now
we just store the data in 'struct in_addr' format.

This also reduce the memory used by Monkey as the previous model holds
in memory buffers for each possible incomming network address. The
memory usage is reduced in about 20KB.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>