New TransportLayer directive in monkey.conf
This new directive provides to each [SERVER] section the capability to
define which transport layer the core server will use. The value of the
key is the shortname of one of the network I/O plugin defined in the
configuration file plugins.load.

Please check the following example for the file monkey.conf:

To provide just HTTP use:

    TransportLayer liana

To provide SSL over HTTP (https):

    TransportLayer liana_ssl

Signed-off-by: Eduardo Silva <edsiper@gmail.com>