Core: use epoll LEVEL_TRIGGERED mode
Monkey events handler was using the EDGE_TRIGGERED (EPOLLET) for the
sockets file descriptors, also doing an unnecessary ioctl() to FIONREAD,
this was causing an extra load which can be avoided with LEVEL_TRIGGERED
and doing some better handling of buffers to read() incoming data.

This patch also fix an unset new_size counter when reallocating the read
buffer.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>