Fix double-free() on client_session when keep-alive is off
When KeepAlive feature is set to off, Monkey was doubling free
the client_session struct, this was done in mk_http_request_end()
and also in mk_conn_write().

The patch removes the explicit free in mk_conn_write() and let
return mk_http_request_end() the final status.

This issue was discovered and initially fixed with a patch
provided by "hio_ <trustthesky@gmail.com>". This new patch
handles the conditionals in a different way.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>