Socket: mk_socket_ip_str() accepts input buffer
The function which format IP strings was allocating memory buffer
for the string, this patch makes the function to accept an input
buffer so the caller must care about it.

Also this patch modifies the logger plugin to use the new API of
the function in question. Now it uses a thread key to hold the
buffer in memory used to format the IP string, it should be
a little bit faster now.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>