Duda: new response API object
the response API object expose methods to configure and perform
a response of data to the client, this patch implements the following
methods:

  - http_status: set the response HTTP status
  - http_header: add a new HTTP response header
  - body_write : write content in the body part of the response
  - end        : finalize the work flushing data as required

This patch also implements the body_buffer concept. Each duda_request_t
have associated a 'struct iovec *' array which is used by body_write()
to perform socket writes. These writes are asynchronous and done when
the callback is ending, leaving Duda and Monkey Core to handle the
output data.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>