Duda: body_buffer adjust to asynchronous mode (event oriented)
When writting the body_buffer, due to the asynchronous sockets mode,
less data than the expected can be written, so Duda must catch up this
behavior.

This patch makes to adjust an offset in the body_buffer, as well it
register a callback with the write events triggered by the core of
Monkey, so in the next event calls it sends the pending data.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>