Duda: new packages support!
This patch adds a new packages mechanism, so web services can
request to Duda to load specific packages/objects.

The first package available with Duda is JSON, so to add JSON
support to a web service, it will take two steps:

  #include "packages/json/json.h"

  int duda_init(struct duda_api_objects *api)
  {
      ...
      duda_service_init();
      duda_load_package(json, "json");
      ...
  }

The duda_load_package() call takes two arguments, the first one
is the package object which expose the API calls, the second is
the name of the package.

The packages lookup takes place in duda.conf file under the configuration
key 'PackagesRoot'.

Now the fun begins :)

Signed-off-by: Eduardo Silva <edsiper@gmail.com>