Request: fix memory leak when invoking premature close
When premature close was invoked, it requested memory for a temporal
session_request. On this condition that memory block was never freed.

The patch drops the mk_mem_alloc() call and just link the temporal
session_request pointer to the sr_fixed field from the client_session
struct.

Thanks to Felipe Astroza for report this issue!

Signed-off-by: Eduardo Silva <edsiper@gmail.com>