Scheduler: optimize connection queues
This patch does a great improvement in terms of performance. When
a write event is triggered the scheduler is invoked to verify if the
connection already exist or if it refers to a new one.

The past scheduler implementation had a fixed array per thread with
slots for a number of (sched_connection struct * worker_capacity). The
main problem was that when many connections were active, do a sequencial
search was very expensive due to the fragmentation originated by the
'status' field of the the sched_connection entries in the whole array.

In the mailing list, developers were discussing a better approach to
register a new connection and Max mentioned that having two queues of
connection could improve something and be more clean.

This patch applies the "two queues" idea proposed to the scheduler but
improving not only the connection register, also improving the hard lookup
when the write event is triggered, the new mechanism is as follows:

   struct sched_connection {
      ...
      mk_list busy_queue;    /* active sched connections */
      mk_list av_queue;      /* available sched connections */
      ...
   };

Comparing and profiling with valgrind/callgrind we get the following
performance results:

         Function name               Inclusive Cost    Self Cost
         -------------------------------------------------------
 old ->  mk_sched_get_connection()       10.63          10.63
 new ->  mk_sched_get_connection()        0.86           0.86

The new mechanism optimize the scheduler connections lookup in a 90%!,
conclusion?, good public discussions get great ideas :)

Signed-off-by: Eduardo Silva <edsiper@gmail.com>