Duda: params: new get_number() method
This new method add the capability to get a converted parameter
from it string origin to a long data type, the work is done without
extra memory allocations and it take cares of long overflows. The
implementation its almost equal to strtol() but it uses a fixed
base 10 and accept an offset for the string length.

The patch also add a new file duda_utils.c and duda_utils.h where
the duda_utils_strtol() function lives.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>