Configure: fix #155 new accept() mode in configure flags
This patch add two new flags to the configure script:

    --only-accept
    --only-accept4

These flags let specify the builder system which accept mode
must be used in the server socket: accept() or accept4(). If
no one is specified the builder will auto-test accept4(), if
it fails, it will go for accept().

This is useful when doing cross compiling and want to specify a
specific operation mode.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>