Scheduler: register new incoming sockets with MK_EPOLL_READ
Previous this patch, the Scheduler used to register each new incoming
socket with the MK_EPOLL_WRITE flag, so when the connection was established
epoll_wait(2) triggered the write event, meaning 'the socket is ready to
write'. On that moment the Connection handler realised that no Scheduler
information exists for that socket and run the mk_sched_register_client()
function, then invoke epoll_ctl(2) to change the socket mode to
MK_EPOLL_READ, so it start to wait for the incoming request string.

In HTTP the client always begins the communication sending the request, so
having a socket on MK_EPOLL_WRITE and then change it to MK_EPOLL_READ adds
a little overhead.

This patch makes the initial socket be registered with MK_EPOLL_READ and
make the Connection handler identify this new connection in the
mk_conn_read() function, avoiding the extra system call described in the
previous behavior.

Here is a little benchmark done with Valgrind/Callgrind:

  +------------------------+----------------+-----------------+--------+
  | function()             | before patch   |  after patch    |  gain  |
  +------------------------+------+---------+-----------------+--------+
  |                        | self | called  |  self | called  |        |
  +------------------------+------+---------+-------+---------+--------+
  | epoll_ctl(2)           | 0.15 |  15000  |  0.10 | 10000   |  44%   |
  +------------------------+------+---------+-------+---------+--------+
  | mk_epoll_change_mode() | 0.57 |  10000  |  0.27 |  5000   |  63%   |
  +------------------------+------+---------+-------+---------+--------+

Signed-off-by: Eduardo Silva <edsiper@gmail.com>