Request: fix possible buffer overflow in Host header
The Host header allow the input of the hostname plus the TCP
port used, e.g:

 GET / HTTP/1.1
 Host: localhost:2001

If the port length was higher than 6 bytes, it make monkey crash
due an overflow in the port buffer.

This patch restrict the port length and also add a validation
to the port value so it just can hold numeric values.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>