epoll: New epoll states
The Monkey API allows third party components to hook file descriptors
into the main worker epoll loop and base on the events it triggers the
proper callbacks through the plugins interface, but Monkey did not
track the epoll mode flags as this was just known by the Kernel.

This patch introduces the 'epoll states' mechanism which keeps track
of every file descriptor in the epoll loop into a user-space linked
list per worker using thread keys. In addition, this feature is the
base for new epoll modes implemented in the core: MK_EPOLL_SLEEP and
MK_EPOLL_WAKEUP.

From now the core and plugins can set an epoll file descriptor in
sleep mode (do not trigger events) and wakeup mode to restore events
to the state before the sleep mode.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>