dirlisting: Fix failure in big directories
Dirlisting is not designed to support event-based working, it is one-shot.

So in a big directory, writev would return EAGAIN (EWOULDBLOCK) after
having written a partial entry.

This would then confuse the client in chunked mode (you promise X bytes but
only send less), which closes the connection, leaving you with a partial listing.

With this patch, dirlisting succeeds in a dir with 23k entries.
Without, it fails in a dir with little over 1k entries.

Signed-off-by: Lauri Kasanen <cand@gmx.com>
Signed-off-by: Eduardo Silva <edsiper@gmail.com>