Signals: implement sys_siglist if symbol is not found
When compiling with old toolchains for some ARM devices and
mostly related to old versions of uClib, the symbol sys_siglist
is not found and make the compilation process to fail.

This patch introduce a header with the sys_siglist array defined
in the following file:

  src/contrib/uclib/sys_siglist.h

If the symbol is not found, it includes the header.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>