Headers: Fix #170: sr->headers._extra_rows keeps invalid reference
Found a specific condition in Monkey core where after send
the headers, if the field sr->headers._extra_rows is set and
then freed, its not marked as NULL. If sub sequent calls
from the Plugin API takes place and this field is validated,
it fails in a segmentation fault.

This patch set the structure field to NULL after is freed.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>