lib: do not left stage30 until all content is sent
If for some reason the send() call returns -1 because the socket
is not yet ready to receive more data (due to non-blocking mode),
the routine should not stop working and should try to send the
information over and over until is fully served.

This is not an expected behavior in a non-blocking server, but is
a temporal fix until the 'pending buffers' interface is implemented.

This problem was found due to #170.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>