Liana_SSL: deprecate plugin based on MatrixSSL
According to:

  http://lists.monkey-project.com/pipermail/monkey/2013-February/001785.html

Liana provides the networking layer to Monkey and its _SSL version the
networking layer with SSL encryption capabilities. The first version of
this plugin was built on top of MatrixSSL library but this last one is not
well maintained and there is no community involved on that product, what
that means ? that nobody cares about it and integrate in a Linux
distribution is a pain so we cannot be stuck on that anymore. Jonathan
(zeus) from our community was doing a good job packaging the latest version
in our repository, but MatrixSSL continue be a problem.

After review some open sourced SSL options, I found that PolarSSL is the
best one that fit our needs. Because of:

   - open source
   - well maintained in Linux distributions
   - lightweight so it fill well in embedded Linux environments
   - years of work supported by a community

Besides that, there is a concern about how that library can impact Duda I/O
(web services framework that runs on top of Monkey). Because the web
services created by people can be closed...  what happen if the service
links to PolarSSL through Duda LGPL code ? answer: nothing. As PolarSSL
works under Monkey and Monkey under Duda, there is no problem in linking
or licensing issues, this have been confirmed by PolarSSL team after talk
to their lawyer, so we are pretty good.

Sonny Karlsson have provided yesterday a new plugin named PolarSSL (good
job!), so everything is ok for the next step.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>