Fix #182: DoS bug on headers parser.
This patch fix the root cause for a problem described in Ticket #182,
actually if a header is malformed like a Header Key without a value, the
ToC parser used to continue processing the next header line.

The solution applied is to improve the ToC generator where it adds extra
validations for at least one colon and forcing each header line to contain
a value or empty space, otherwise the server will trigger a Bad Request
response to the client and close the connection.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>