auth: Fix buffer overflow when loading user file.
The third parameter of strncpy(3) was badly calculated.
For each entry in the users file the parameter was the length from start
of file until end of username, when it should be calculated from current
buffer position.
This causes writes outside buffer when the user file is larger than 2 *
sizeof(struct user).

Signed-off-by: Sonny Karlsson <ksonny@lotrax.org>