Mimetype: replace bsearch by red-black-tree (50% improvement)
The mime types lookup algorithm was using a qsort()ed array plus a
binary search with two main arrays: a top ten list (common) and another
one for the rest.

Even the solution was optimal for some cases, it have a huge dependency
of a righ configuration of the top ten list, and considering the support
of virtual host and different types of content we need a more generic
solution that scale and optimize for all cases.

This patch deprecate the top-ten list and replace the binary search by a
red-black-tree solution, looking at the results after a benchmark we can
see the following:

                       +-----------+-----------------+--------+
                       | bsearch   | red-black-tree  |   gain |
 +---------------------+-----------+-----------------+--------+
 |  mk_mimetype_find   |   11.54   |           5.82  |   ~50% |
 +---------------------+-----------+-----------------+--------+

More graphical details here:

 http://edsiper.linuxchile.cl/mimetype_performance_improvement_001.png

Signed-off-by: Eduardo Silva <edsiper@gmail.com>