Scheduler: Workaround for ULONG_MAX bug.
This is a workaround for a specific Bug that we still not find the
root cause for it. The conditions are:

  - for some reason closed_connections > active_connections
  - when performing (active_connections - closed_connections it will
    return UMAX_LONG, which is OK but *not* for our case

  - if cur (current connections) is greater than worker capacity it will
    fail all new incoming connections, and as it may have ULONG_MAX that
    would case a very very bad behavior

The temporal workaround is to when facing that we are over capacity,
check if we have a ULONG_MAX value in 'cur', on that moment signal
our workers so they can start performing a synchronization of their
counters.

The patch adds a signal channel for each worker so every epoll loop
will monitor that channel, on receive the 0xDEADBEEF value, it will
perform a synchronization of counters based on scheduler busy queue.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>