Error Handling: new macros mk_libc_error() and mk_libc_warning()
This patch adds two new functions and proper macros to handle common
'errno' errors and format them to stdout.

The new static inline functions are located in src/include/utils.h:

   mk_utils_libc_error(char *file, int line);
   mk_utils_libc_warning(char *file, int line);

the wrappers (based on macros) to be use are:

   mk_libc_error();
   mk_libc_warning();

when used the following information is formatted to stdout: errno number,
errno message, file and line.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>