Socket: handle new Kernel 3.14 tcp_autocorking feature.
the server use the TCP_CORK feature to instruct the Kernel to
wait for some data before dispatch it on the socket buffer. This
is a common approach when sending HTTP response headers and then
a content body (file data).

The Linux Kernel 3.14 introduce a new feature called 'tcp autocorking',
it basically do TCP_CORK by default based on specific conditions of
each socket status. This is a good approach to let the Kernel to decide
when it can be used and also it avoid user space to invoke two system
calls to enable/disable the TCP_CORK.

This patch makes the Server to validate if the new feature exists on the
running Kernel and if is enabled, it will disable all manual corking done,
for more details about how it works at Kernel level refer to the following
patch:

    http://goo.gl/AMFIuI

Note: on an older Kernel it will continue working as usual.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>