Scheduler: optimization when timing out connections
This patch adds a good improvement on performance when the event poll
system start checking for candidates to timed out connections. It adds
two main lists:

  - sched->status_queue: Open socket connections list that have not sent any data
  - cs_incomplete: Client requests list that still have some pending data

When the timer triggers the timeout check, now the Scheduler will look
over those two global lists at thread level, avoiding to check connections
that are in a good state.

This performance improvement is faced when having several active connections
in the server.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>