Epoll: validate when registering connection with the Scheduler
In the epoll event loop, when registering a new connection with
the Scheduler, there were not check if this was successfull or not,
a common example may be a security plugin dropping a connection
at STAGE_10.

This patch adds a validation on the return value when trying to register
a connection with the Scheduler. If it fails, the socket will be closed
right away.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>