Events: change flags value and found a libkqueue bug
I found a bug running libkqueue (GIT version), not sure if it
can be replicated on older versions. It only happens when the
following sequence of kqueue calls are done:

 1. Register FD as EVFILT_READ
 2. Delete EVFILT_READ for FD
 3. Register FD as EVFILT_WRITE
 4. Register FD as EVFILT_READ

On step #4 kevent() fails with EFAULT (Bad Address). I was able
to reproduce the problem with a standalone test case, I contacted
the library author about this problem. I need this to get resolved
to continue.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>