Events: new mk_event_translate() and mk_event_foreach() interfaces
Previous mechanism required to perform two walks through the events
result array, the first one for the native backend and the second
to export the results to a new interface.

This patch implements two new interface to improve performance over
the past scenario plus a minor change:

1. mk_event_wait(): just invoke the backend poll interface, nothing else.
2. mk_event_foreach() macros that allows to walk arround over the backend
   interface results avoiding double copy.
3. mk_event_translate(): do a explicit copy from backend resutls to the
   public interface.

This patch only apply the interfaces to the epoll backend.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>