Core: Plugin: Let build system hook static plugins
This patch implements a new option on the configure script called 'static_plugins='
where is possible to define which plugins will be build in static mode. Also the
core is able to detect the statics and link them in the global configuration.

More work is required, but this is a good progress. Just tested on Liana plugin.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>