HTTP: restore performance on Linux for small static files.
Later on 2014 Monkey performance got decreased when serving few bytes
of data on Linux. A fix for OSX generated a performance issue on Linux
that is only faced when mixing sendfile(2) and TCP_CORK.

This patch implements the fix for Linux, now it increased performance
up to ~24% when serving small static files. Monkey is rocking again :)

Signed-off-by: Eduardo Silva <eduardo@monkey.io>