Core: support new PTHREAD_TLS mode
Monkey support threads, and when sharing global data on each worker it does
it by using the Compiler TLS (Thread Local Storage) feature.

TLS is a compiler specific feature, it works on GCC and CLANG. As it can be
considered a *new* feature on compilers, some legacy compiler versions or
custom toolchains for Embedded do not come with that feature.

This patch adds a new option called PTHREAD_TLS that can be enabled on the
configure script through the --pthread-tls option. When enabled, instead
of TLS, Monkey will use Pthread Keys for the same purpose.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>