Core: Check Listeners at start (Fix #173).
This patch implements a new mechanism to test if a Listener interface and
TCP port are already in use through a basic TCP connection. This check
is required as the sockets can be in shared mode (SO_REUSEPORT) we need
to be aware if other Monkey instance is already running.

If the test claims there is an interface busy, it will abort. But we give
the user a new option '-T' to override this failure and proceed anyways.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>