Issue 170 implemented : Syslog functional and optional
Syslog has been implemented and can be made to work using the --enable-syslog option during configuration.
This option defines Syslog in mk_info.h file which in turn is read by logger.c

At present logger facility on master is broken so Syslog will not work even though it has been implemented. As soon as logger is fixed it will start working too.
Syslog utilizes syslog.h header file and logs to /var/log/syslog

Now based on whether the option is set or not, logger.c decided whether to log to syslog or to the log files()