Build: start migration to CMake (yes, I changed my mind)
This patch adds the first changes of the migration from the old
bash build system to a compatible CMake.

This is still work in process and it aims to keep a generic 'configure'
script with the same options than before but it uses CMake in the
backend to validate dependencies and generate the Makefiles.

I will explain later why I changed my mind about this in blog post, for
now there is a lot of work to do. :)

Signed-off-by: Eduardo Silva <eduardo@monkey.io>