HTTP: Fix memory corruption when handling a protocol error (Fix #194)
When the HTTP parser catch an error, the HTTP core will try to close(2) an
unitialized file descriptor from the session, generating a memory corruption.

This patch make sure to initialize the value and only try to close it if
it have some valid number.

note: this problem is only faced on v1.6.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>