mk_core: event (mk_server): use event context when deleting a connection
this patch makes the event handler to use the connection context mk_event as
a reference to obtain the connection file descriptor and event bitmask.

The main reason for this change is the requirement to get the existent bitmask for
the kqueue interface, when deleting an event it needs the 'filter' associated and this
is translated to the event mask: EVFILT_READ ; EVFILT_WRITE.

For epoll(2) backend is not required but there is no performance penalty associated,
epoll backend patch already coming.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>