mk_core: implement a static buffer for libc error messages (fix #195)
Monkey utils implements a buffer per thread to format libc errors, but
when the mk_utils_libc_error() is invoked from the parent process, it
may crash as the pthread key do not exists.

This patch implements a static buffer for those cases when the call is
invoked from a parent process.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>