Plugins: FastCGI: new FastCGI plugin
This new implementation of the fastcgi interface use the new Monkey
architecture and reduce the code complexity due to the new features
available on the server core.

It drops some neat features from the old version as KeepAlive and
the ability of multiple connections, all that will be restored in the
new version of Monkey which will support server pools natively.

Note: old plugin required like 3000 lines of code, the new one only
around 650.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>