mk_server: add handlers support
this patch enable a new Handlers support for plugins, so now is
possible to define custom handlers for specific URI matches based
on regular expressions, a virtual host may enable FastCGI through
the following setup:

  [HANDLERS]
      Match /.*\.php fastcgi

The 'Match' rule indicate that it's about a regular expression, the
second argument is the regexec and the third argument is the plugin
handler name (known as short name).

Handlers are processed in order, if no handler own the request the
server will dispatch the file content.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>