Plugins: fastcgi: fix crash when Jemalloc is enabled (Fix #221)
When Monkey is built with Jemalloc, when the plugin release the resources
associated to a FastCGI request, it crash. Jemalloc reports the following
error:

  Failed assertion: "pageind >= map_bias"

This problem happens because when the plugin compose the FastCGI request
internally, it needs to duplicate some string buffers and it was using
a direct strdup(3) call, which is wrong as Monkey provides wrappers for
memory and string manipulation which use the system allocator (malloc)
or the jemalloc functions. The problem was:

 - Duplicate string buffer with strdup(3)
 - Release buffer with Jemalloc

This patch fix the single line that generated the problem.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>