mk_core: event: new select(2) backend
This patch implement a new event poll mechanism backend based on
the select(2) interface. This is done to provide some compatibility
across other operating systems.

Due to the native limitations of select(2), it's not recommended
to handle a high number of concurrent file descriptors. The limit is
usually imposed by FD_SETSIZE (1024).

Signed-off-by: Eduardo Silva <eduardo@monkey.io>