Monkey: RFC1867 implementation for single file upload
The fact is that today Monkey caches the whole HTTP request body in the
process heap (realloced) for any kind of file upload case and it passes
the file to plugins like FCGI etc. in several chunks, only after
complete file is downloaded within Monkey's heap area. Now for bigger
files (in hundreds on MBs or GBs) or for systems with low RAM, this kind
of model is actually inefficient. Not only it stresses the process heap
and the memory manager (due to constant realloc), the real space
(RAM/flash) requirement is actually almost double the file size (because
the applications like FCGI also creates separate copy of the same file
either in RAM file system or in flash) and unless the entire file is
sent to plugins, Monkey doesn't free the file copy kept in the process
heap. This causes the low RAM embedded systems to go for OOM for larger
file upload scenarios.

This can be solved in 2 ways:

1. Either implement a pass through mechanism in Monkey and if seen file
upload request, keep passing the request body data to plugin
applications (FCGI etc.) as and when data arrives and let the plugins
reassemble all the parts of the data in file system and rebuild the file
back. Something very similar is proposed here: #228. The disadvantage of
this model is that we might compromise on some security aspects. Also
this means we expect the plugins to be HTTP protocol aware.

2. The alternate model is to directly cache the file data to file system
instead of process heap and once the whole request body is downloaded,
verify the RFC1867 parameters and then pass the file path and the form
attributes to FCGI.

The attached patch tries to implement single file upload using
'multipart/form-data' as described in RFC1867. In future we can
generalize the patch to support multiple file upload in same form using
'multipart/mixed'. But at this moment my requirement is limited to
single file upload. Hence I would request to kindly review the attached
patch file for the same implementation and let me know your comments to
improve it. Also the patch is limited to my immediate required platform
Linux. However the logic remains generic and we can easily add support
for other platforms whenever required.